agent/requirements.txt
python-dotenv
livekit-agents>=1.4.3
livekit-plugins-browser>=0.1.4
livekit-plugins-anthropic>=1.4.3
